refactor 'database' to scans_database
